Move Event class to util folder
